### 2022-04-10 #4308 @editorialbot
[PRE REVIEW]: EPICpy: A Python Interface for EPIC Computational Cognitive Modeling
**Submitting author:** @travisseymour (<a href="http://orcid.org/0000-0003-4719-9152">Travis Seymour</a>)
**Repository:** https://github.com/travisseymour/EPICpy
**Branch with paper.md** (empty if default branch): JOSSArticle
**Version:** v2022.3.5
