### 2022-04-10 49b860a
more emoji_box changes -- moved things to be more vertical than horizontal to accommodate thinner output windows.
