+++ src/main/python/epicsimulation.py
@@ -163,7 +163,8 @@ class Simulation:
         except Exception as e:
             self.write(
                 emoji_box(
-                    f"ERROR: Failed to update EPIC output or trace settings: {e}",
+                    f"ERROR: Failed to update EPIC output or trace settings:\n"
+                    f"{e}",
                     line="thick",
                 )
             )
@@ -269,10 +270,10 @@ class Simulation:
                 except IOError:
                     self.write(
                         emoji_box(
-                            f"ERROR: Unable to create needed file: "
-                            f'"{str(Path(device_file_root, "__init__.py"))}". If your '
-                            f"device may not load correctly, you may need to obtain "
-                            f"write access to the device folder.",
+                            f'ERROR: Unable to create needed file \n'
+                            f'"{str(Path(device_file_root, "__init__.py"))}".\n'
+                            f'If your device does not load correctly, you may need to\n'
+                            f'obtain write access to the device folder.',
                             line="thick",
                         )
                     )
@@ -319,8 +320,9 @@ class Simulation:
             except AssertionError as e:
                 self.write(
                     emoji_box(
-                        f"ERROR: Failed to create new EpicDevice from "
-                        f"{device_file_p.name}! [{e}]",
+                        f"ERROR: Failed to create new EpicDevice from\n"
+                        f"{device_file_p.name}!\n"
+                        f"{e}",
                         line="thick",
                     )
                 )
@@ -328,8 +330,9 @@ class Simulation:
                 return
             except IOError as e:
                 self.write(
-                    f"ERROR: Unable to write to file "
-                    f"{str(self.tempmod_path.parent)} ({e})."
+                    f"ERROR: Unable to write to file\n"
+                    f"{str(self.tempmod_path.parent)}:\n"
+                    f"{e}"
                 )
             except Exception as e:
                 log.error(
@@ -389,8 +392,9 @@ class Simulation:
             except Exception as e:
                 self.write(
                     emoji_box(
-                        f"ERROR: Failed to create new EpicDevice from "
-                        f"{device_file_p.name}! [{e}]",
+                        f"ERROR: Failed to create new EpicDevice from\n"
+                        f"{device_file_p.name}!\n"
+                        f"{e}",
                         line="thick",
                     )
                 )
@@ -450,7 +454,8 @@ class Simulation:
             except Exception as e:
                 self.write(
                     emoji_box(
-                        f"ERROR: Unable to properly connect new device and model [{e}]",
+                        f"ERROR: Unable to properly connect new device and model:\n"
+                        f"{e}",
                         line="thick",
                     )
                 )
@@ -540,8 +545,9 @@ class Simulation:
         except Exception as e:
             self.write(
                 emoji_box(
-                    f"ERROR: Unable to unload {kind} encoder! [{e}] However, it won't "
-                    f"be loaded the next time this device is loaded.",
+                    f"ERROR: Unable to unload {kind} encoder!\n"
+                    f"[{e}]\n"
+                    f"However, it won't be loaded the next time this device is loaded.",
                     line="thick",
                 )
             )
@@ -606,8 +612,9 @@ class Simulation:
             except Exception as e:
                 self.write(
                     emoji_box(
-                        f"ERROR: Failed to create new {kind} encoder from "
-                        f"{encoder_file_p.name}! [{e}]",
+                        f"ERROR: Failed to create new {kind} encoder from\n"
+                        f"{encoder_file_p.name}:\n"
+                        f"{e}",
                         line="thick",
                     )
                 )
@@ -660,8 +667,9 @@ class Simulation:
             except Exception as e:
                 self.write(
[…]
