### 2022-04-13 7a4bdc2
added full complement of directional arrows to shape router. No new shpae code needed, just had to specify angular offsets.
