+++ src/main/python/views/visualviewwindow.py
@@ -441,6 +441,14 @@ class VisualViewWin(QMainWindow):
             "Down_Arrow": partial(self.shape_right_arrow, rotation=90),
             "Left_Arrow": partial(self.shape_right_arrow, rotation=180),
             "Right_Arrow": partial(self.shape_right_arrow, rotation=0),
+            "NN_Arrow": partial(self.shape_right_arrow, rotation=-90),
+            "SS_Arrow": partial(self.shape_right_arrow, rotation=90),
+            "WW_Arrow": partial(self.shape_right_arrow, rotation=180),
+            "EE_Arrow": partial(self.shape_right_arrow, rotation=0),
+            "NE_Arrow": partial(self.shape_right_arrow, rotation=-45),
+            "NW_Arrow": partial(self.shape_right_arrow, rotation=225),
+            "SE_Arrow": partial(self.shape_right_arrow, rotation=45),
+            "SW_Arrow": partial(self.shape_right_arrow, rotation=135),
             "Cursor_Arrow": partial(self.shape_right_arrow, rotation=240),
             "Line": self.shape_line,
             "T000": self.shape_t_object,
