### 2022-04-13 203873f
removed forgotten debug logs from config module
