### 2022-04-13 b74b78b
reclassified debug logs in cpyysetup to info logs
