### 2022-04-13 b23489c
repositioned Quit on NormOutWin context popup to prevent accidental exits.
