### 2022-04-13 17ff0e2
Improved search dialog. backwards button plus enter to close.
