+++ src/main/python/uifiles/searchui.py
@@ -16,26 +16,26 @@ from PySide2.QtWidgets import *
 class Ui_DialogSearch(object):
     def setupUi(self, DialogSearch):
         if not DialogSearch.objectName():
-            DialogSearch.setObjectName("DialogSearch")
+            DialogSearch.setObjectName(u"DialogSearch")
         DialogSearch.setWindowModality(Qt.ApplicationModal)
         DialogSearch.resize(702, 245)
         DialogSearch.setMinimumSize(QSize(702, 242))
         DialogSearch.setMaximumSize(QSize(3000, 3000))
         font = QFont()
-        font.setFamily("Consolas")
+        font.setFamily(u"Consolas")
         font.setPointSize(14)
         DialogSearch.setFont(font)
         DialogSearch.setModal(True)
         self.gridLayout = QGridLayout(DialogSearch)
-        self.gridLayout.setObjectName("gridLayout")
+        self.gridLayout.setObjectName(u"gridLayout")
         self.verticalLayout_2 = QVBoxLayout()
-        self.verticalLayout_2.setObjectName("verticalLayout_2")
+        self.verticalLayout_2.setObjectName(u"verticalLayout_2")
         self.verticalLayout = QVBoxLayout()
-        self.verticalLayout.setObjectName("verticalLayout")
+        self.verticalLayout.setObjectName(u"verticalLayout")
         self.label = QLabel(DialogSearch)
-        self.label.setObjectName("label")
+        self.label.setObjectName(u"label")
         font1 = QFont()
-        font1.setFamily("Consolas")
+        font1.setFamily(u"Consolas")
         font1.setPointSize(14)
         font1.setBold(True)
         font1.setWeight(75)
@@ -43,81 +43,76 @@ class Ui_DialogSearch(object):
 
         self.verticalLayout.addWidget(self.label)
 
-        self.plainTextEditSearchText = QPlainTextEdit(DialogSearch)
-        self.plainTextEditSearchText.setObjectName("plainTextEditSearchText")
-        self.plainTextEditSearchText.setFont(font)
-        self.plainTextEditSearchText.setStyleSheet("")
+        self.lineEditSearchText = QLineEdit(DialogSearch)
+        self.lineEditSearchText.setObjectName(u"lineEditSearchText")
+        self.lineEditSearchText.setClearButtonEnabled(True)
 
-        self.verticalLayout.addWidget(self.plainTextEditSearchText)
+        self.verticalLayout.addWidget(self.lineEditSearchText)
+
 
         self.verticalLayout_2.addLayout(self.verticalLayout)
 
         self.checkBoxRegEx = QCheckBox(DialogSearch)
-        self.checkBoxRegEx.setObjectName("checkBoxRegEx")
+        self.checkBoxRegEx.setObjectName(u"checkBoxRegEx")
         self.checkBoxRegEx.setFont(font1)
 
         self.verticalLayout_2.addWidget(self.checkBoxRegEx)
 
         self.checkBoxIgnoreCase = QCheckBox(DialogSearch)
-        self.checkBoxIgnoreCase.setObjectName("checkBoxIgnoreCase")
+        self.checkBoxIgnoreCase.setObjectName(u"checkBoxIgnoreCase")
         self.checkBoxIgnoreCase.setFont(font1)
 
         self.verticalLayout_2.addWidget(self.checkBoxIgnoreCase)
 
         self.horizontalLayout = QHBoxLayout()
-        self.horizontalLayout.setObjectName("horizontalLayout")
-        self.horizontalSpacer = QSpacerItem(
-            388, 20, QSizePolicy.Expanding, QSizePolicy.Minimum
-        )
+        self.horizontalLayout.setObjectName(u"horizontalLayout")
[…]
