### 2022-04-16 2aed6d9
added ability to draw full cardinal set of leader positions. Shortened leader length slightly.
