### 2022-04-20 ad10da3
added way to delete datafile without opening the run settings menu.
