+++ src/main/python/uifiles/mainui.py
@@ -18,199 +18,197 @@ from cachedplaintextedit import CachedPlainTextEdit
 class Ui_MainWindow(object):
     def setupUi(self, MainWindow):
         if not MainWindow.objectName():
-            MainWindow.setObjectName("MainWindow")
+            MainWindow.setObjectName(u"MainWindow")
         MainWindow.resize(638, 640)
         font = QFont()
-        font.setFamily("Consolas")
+        font.setFamily(u"Consolas")
         font.setPointSize(14)
         font.setBold(True)
         font.setWeight(75)
         MainWindow.setFont(font)
         self.actionLoad_Device = QAction(MainWindow)
-        self.actionLoad_Device.setObjectName("actionLoad_Device")
+        self.actionLoad_Device.setObjectName(u"actionLoad_Device")
         font1 = QFont()
-        font1.setFamily("Consolas")
+        font1.setFamily(u"Consolas")
         font1.setPointSize(12)
         font1.setBold(True)
         font1.setWeight(75)
         self.actionLoad_Device.setFont(font1)
         self.actionCompile_Rules = QAction(MainWindow)
-        self.actionCompile_Rules.setObjectName("actionCompile_Rules")
+        self.actionCompile_Rules.setObjectName(u"actionCompile_Rules")
         self.actionCompile_Rules.setEnabled(False)
         self.actionCompile_Rules.setFont(font1)
         self.actionRecompile_Rules = QAction(MainWindow)
-        self.actionRecompile_Rules.setObjectName("actionRecompile_Rules")
+        self.actionRecompile_Rules.setObjectName(u"actionRecompile_Rules")
         self.actionRecompile_Rules.setEnabled(False)
         font2 = QFont()
-        font2.setFamily("Consolas")
+        font2.setFamily(u"Consolas")
         font2.setPointSize(12)
         self.actionRecompile_Rules.setFont(font2)
         self.actionQuit = QAction(MainWindow)
-        self.actionQuit.setObjectName("actionQuit")
+        self.actionQuit.setObjectName(u"actionQuit")
         self.actionQuit.setFont(font1)
         self.actionRun_Settings = QAction(MainWindow)
-        self.actionRun_Settings.setObjectName("actionRun_Settings")
+        self.actionRun_Settings.setObjectName(u"actionRun_Settings")
         self.actionRun_Settings.setEnabled(False)
         self.actionRun_Settings.setFont(font2)
         self.actionDisplay_Controls = QAction(MainWindow)
-        self.actionDisplay_Controls.setObjectName("actionDisplay_Controls")
+        self.actionDisplay_Controls.setObjectName(u"actionDisplay_Controls")
         self.actionDisplay_Controls.setEnabled(False)
         font3 = QFont()
-        font3.setFamily("Consolas")
+        font3.setFamily(u"Consolas")
         font3.setPointSize(12)
         font3.setBold(False)
         font3.setWeight(50)
         self.actionDisplay_Controls.setFont(font3)
         self.actionTrace_Settings = QAction(MainWindow)
-        self.actionTrace_Settings.setObjectName("actionTrace_Settings")
+        self.actionTrace_Settings.setObjectName(u"actionTrace_Settings")
         self.actionTrace_Settings.setEnabled(False)
         self.actionTrace_Settings.setFont(font3)
         self.actionRule_Break_Settings = QAction(MainWindow)
-        self.actionRule_Break_Settings.setObjectName("actionRule_Break_Settings")
+        self.actionRule_Break_Settings.setObjectName(u"actionRule_Break_Settings")
         self.actionRule_Break_Settings.setEnabled(False)
[…]
