### 2022-04-21 f80b87e
Normalized font for new datafile delete menu item. Also added dynamic enabling/disabling depending on sim run state.
