+++ src/main/python/uifiles/mainui.py
@@ -201,18 +201,22 @@ class Ui_MainWindow(object):
         self.actionAll_Runs.setFont(font2)
         self.actionDelete_Datafile = QAction(MainWindow)
         self.actionDelete_Datafile.setObjectName(u"actionDelete_Datafile")
+        font4 = QFont()
+        font4.setFamily(u"Consolas")
+        font4.setPointSize(14)
+        self.actionDelete_Datafile.setFont(font4)
         self.centralwidget = QWidget(MainWindow)
         self.centralwidget.setObjectName(u"centralwidget")
         self.centralwidget.setFont(font3)
         self.plainTextEditOutput = CachedPlainTextEdit(self.centralwidget)
         self.plainTextEditOutput.setObjectName(u"plainTextEditOutput")
         self.plainTextEditOutput.setGeometry(QRect(10, 10, 501, 431))
-        font4 = QFont()
-        font4.setFamily(u"Consolas")
-        font4.setPointSize(14)
-        font4.setBold(False)
-        font4.setWeight(50)
-        self.plainTextEditOutput.setFont(font4)
+        font5 = QFont()
+        font5.setFamily(u"Consolas")
+        font5.setPointSize(14)
+        font5.setBold(False)
+        font5.setWeight(50)
+        self.plainTextEditOutput.setFont(font5)
         self.plainTextEditOutput.setContextMenuPolicy(Qt.CustomContextMenu)
         self.plainTextEditOutput.setReadOnly(True)
         self.plainTextEditOutput.setCenterOnScroll(False)
