### 2022-04-25 4ee3912
Allowing encoders that derrive from new epicpy encoder bases.
