### 2022-05-16 #4308 @travisseymour
Here are some reviewers that seem to have similar interests and Python as
the preferred language:

1. TomDonoghue ***@***.***)
2. drussellmrichie ***@***.***)
3. szorowi1 ***@***.***)

let me know if you need more options.
best
Travis

On Thu, May 12, 2022 at 7:15 PM Prashant K. Jha ***@***.***>
wrote:

> Hi @travisseymour <https://github.com/travisseymour>, could you provide
> the names of a couple of potential reviewers for your submission? You can
> check out this link
> <https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822>
> for the list of reviewers for JOSS.
>
> We need one more reviewer to begin the review.
>
