### 2022-05-18 #4308 @prashjha
Hi @drussellmrichie, would you be interested in reviewing this software article for JOSS? I think this work is in your area of interest. You can learn more about the review process [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
