### 2022-05-18 #4308 @drussellmrichie
Hi. Thanks for the invitation. I might be up for this but I think there are people that would be better-suited for this than me. The author of the [PyACT-R package](https://github.com/jakdot/pyactr), [Jakub Dotlačil](https://github.com/jakdot), would seem to be *perfect*, if they have bandwidth. ;-) Or maybe the co-author of his [book](https://github.com/abrsvn/pyactr-book), [Adrian Brasoveanu](https://github.com/abrsvn). If they are not available, then perhaps I can do this...
