### 2022-06-11 #4308 @abrsvn
I am really sorry, I won't be able to do this for several months. It
sounds like a very valuable library though. Thank you for thinking of
me.

On Fri, Jun 10, 2022 at 9:48 PM Prashant K. Jha
***@***.***> wrote:
>
> Hi @abrsvn, would you have time to review this software and the accompanying software article for JOSS? This may be in your area of work. If you need more information, please feel free to ask. Thanks!
>
