### 2022-06-11 #4308 @drussellmrichie
Just chiming in again to say that I still think someone in the ACT-R community would be really well-suited to review this (and better suited than me). Perhaps Dario Salvucci at Drexel University?
