### 2022-08-15 travisseymour/EPICpy#1 @jakdot
Compile rules
Hi,

Thank you for providing the package EPICpy!

Could you please help me with one issue in compiled rules using EPICpy?

I followed installation instructions on Ubuntu 20.4 from EPICpy documentation.

I could install EPICpy and start it without any issues.

However, when I try to load a device and its corresponding ruleset, following these steps:

```
1. File->Load_Device the Choice Device (devices/choice/choice_device.py).
2. file->Compile_Rules the visual-manual choice ruleset (devices/choice/rules/choicetask_rules_VM.prs).
````

I get this error when compiling rules:
```
ℹ Attempting to compile reloaded ruleset...
*** Error in parsing production rule file: line 15 position 44
(Define Named_location Center_screen 0.0 0.0)
Expecting a numeric value in Named_location
An error was found in parsing production rule file; rules not ready to execute
```

Maybe I am overlooking something obvious. Could you please help me?
