### 2022-08-15 travisseymour/EPICpy#1 @travisseymour
Hi,
I just tried the following

1. Installed a fresh copy of Ubuntu 20.04.4  (You said you used 20.4, but I
think you meant 20.04)
2. Ran updates:
    sudo apt update
    sudo apt upgrade
3. Installed a c++ compiler using:
    sudo apt install g++
4. Downloaded files:
   wget https://people.ucsc.edu/~nogard/software/epicpy/EPICpy.deb
   wget https://people.ucsc.edu/~nogard/software/epicpy/devices.zip
5. Unzipped the device folder:
   unzip devices.zip
6. Installed EPICpy:
    sudo dpkg -i EPICpy.deb
7. Ran EPICpy:
    /opt/EPICpy/EPICpy
8. Within EPICpy:
    a. *Load*ed devices/choice/choice_device.py
    b. *Compile*d devices/choice/rules/choicetask_rules_VM.prs

Everything worked as expected.

Oddly, the error you got points to the dot in the second instance of 0.0 on
line 15. I copy and pasted the line from your error message and it still
worked. I know editing and sharing files between linux, mac, and windows
sometimes leads to encoding issues or hidden characters, but we're both
using the same OS. Would you mind pulling a fresh copy of devices.zip and
trying it again?

Another idea is to comment out the line by pre-pending two forward slashes
and then *manually* retyping in the line yourself. E.g.:

//(Define Named_location Center_screen 0.0 0.0)
(Define Named_location Center_screen 0.0 0.0)

Another idea is to replace the floats with integers, e.g.:

(Define Named_location Center_screen 0 0)

A final suggestion is to just comment out the line altogether. It turns out
that these choice-task rules don't even use this named location, so you
could just comment out the entire line and try again. E.g.:

//(Define Named_location Center_screen 0.0 0.0)

If you make any edits to the rule file, don't forget to save it in your
text editor and then choose FILE->Recompile_Rules within EPICpy.

Given that I'm not able to reproduce the problem, knowing what happens in
these cases might help me figure out what's going wrong for you.

best
Travis
