### 2022-08-16 travisseymour/EPICpy#1 @jakdot
Hi,

Yes, I meant Ubuntu 20.04.

I removed the installed version of EPICpy and the downloaded version of devices and followed all your steps that you give above. I still get the same error. However, when I change the floats for the integers in the crucial line, it works (no error message during compilation). When I comment out the line, it also works.

For now it's fine for me, but if you like to debug this further, I'll be happy to help.
