### 2022-08-16 #4533 @jakdot
Hi @prashjha , done with my review. There was one issue with functionality, see here:

https://github.com/travisseymour/EPICpy/issues/1#issue-1338771489

Since the author provides a workaround in the thread on the issue, I see this as solved (I did not close the issue just yet in case the author wants to debug it further).

As far as I am concerned, the paper is accepted. Let me know if you need more from me.
