### 2022-08-16 travisseymour/EPICpy#1 @travisseymour
Thanks,
That really helps, I'll look into it today.
best
Travis
