### 2022-08-24 #4533 @prashjha
Hi @travisseymour, I will have a look at your draft today, and if there are any suggestions I will share them. Once you incorporate those suggestions and do a tagged release+archive release on zenodo, I will hand your paper to EiC for the final decision. It looks like both reviewers are positive so I will recommend accepting to EiCs.
