### 2022-08-24 travisseymour/EPICpy#1 @travisseymour
Yes, I uploaded a new version of devices.zip to the server that uses
integers rather than floats for named locations. Now when anyone follows
the instructions and downloads the sample devices, they will get the new
version.

On Wed, Aug 24, 2022 at 8:54 AM Prashant K. Jha ***@***.***>
wrote:

> Hi @travisseymour <https://github.com/travisseymour>, is this issue
> addressed?
>
