### 2022-08-24 #4533 @travisseymour
Hi Prashant,
I'm new to all of this and I want to make sure I know what you mean.
Are you suggesting I do a tagged release *using git* like this (if so, is
there anything special you want in the message), or is this part also on
zenodo?

git tag -a v1.0.0 -m "Releasing version v1.0.0"

As for the other stuff you suggested, I've never used zenodo, but I assume
you are referring to this part of the setting-the-software-archive
<https://joss.readthedocs.io/en/latest/editorial_bot.html?highlight=zenodo#setting-the-software-archive>
page:

When a submission is accepted, we ask that the authors to create an archive
(on Zenodo <https://zenodo.org/>, fig*share* <https://figshare.com/>, or
other) and post the archive DOI in the REVIEW issue.

Correct?
thanks,
Travis

On Wed, Aug 24, 2022 at 8:51 AM Prashant K. Jha ***@***.***>
wrote:

> Hi @travisseymour <https://github.com/travisseymour>, I will have a look
> at your draft today, and if there are any suggestions I will share them.
> Once you incorporate those suggestions and do a tagged release+archive
> release on zenodo, I will hand your paper to EiC for the final decision. It
> looks like both reviewers are positive so I will recommend accepting to
> EiCs.
>
