### 2022-08-24 1808d18
Fixed a few typos. Clarified OS versions in 2 of places.
