+++ article/paper.md
@@ -18,7 +18,7 @@ bibliography: paper.bib
 
 # Summary
 
-Executive-Process Interactive Control (EPIC) is a computational cognitive architecture that specifies a theory of human performance and a facility to create behavioral simulations using that theory [@Meyer&Kieras1997a; @Meyer&Kieras1997b]. EPIC compacts decades of psychological theory and findings into a rich set of processors, including those that model sensory, perceptual, motor, cognitive, and memory subsystems of the human mind. EPIC is particularly suited to stimulus-response type human performance tasks including those involving multitasking and task-switching. Although many research articles, technical reports, talks, and conference presentations have been published using EPIC, these have been produced by a relatively small number of researchers. One reason EPIC has not been used more widely is due to its complex design and basis in the C++ programming language. Another challenge is the complete lack of documentation on how to set up and use its development environment. A final limitation is that EPIC is only available for MacOS. None of these issues are a problem for an experienced C++ developer, but they can be significant hurdles for those with little or limited C++ expertise, a state that is exceedingly common in the cognitive science community most likely to be interested in using EPIC. 
+Executive-Process Interactive Control (EPIC) is a computational cognitive architecture that specifies a theory of human performance and a facility to create behavioral simulations using that theory [@Meyer&Kieras1997a; @Meyer&Kieras1997b]. EPIC compacts decades of psychological theory and findings into a rich set of processors, including those that model sensory, perceptual, motor, cognitive, and memory subsystems of the human mind. EPIC is particularly suited to stimulus-response type human performance tasks including those involving multitasking and task-switching. Although many research articles, technical reports, talks, and conference presentations have been published using EPIC, these have been produced by a relatively small number of researchers. One reason EPIC has not been used more widely is due to its complex design and basis in the C++ programming language. Another challenge is the complete lack of documentation on how to set up and use its development environment. A final limitation is that EPIC is only available for MacOS. None of these issues are a problem for an experienced C++ developer, but they can be significant hurdles for those with limited C++ expertise, a state that is exceedingly common in the cognitive science community most likely to be interested in using EPIC. 
 
 Rather than C++, the current approach uses the popular Python programing language. Because Python is now the language used in most introductory programming courses, researchers, their collaborators, and their students are more likely to be familiar with it than with C++. EPICpy not only provides a new Python-based graphical interface for EPIC modeling, but it allows programming simulated tasks, and perceptual models in Python [@EPICpyGithub]. In addition, EPICpy is cross-platform and provides executables for Linux, MacOS, and Windows (via WSL2). It is also possible to use EPICpy with the included Docker file on any operating system. EPICpy is accompanied by rich documentation [@EPICpyDocs] for those wanting to install and run existing simulations, those wishing to create and test new simulations, as well as those wishing to set up an environment for developing EPICpy itself. Hopefully EPICpy and its documentation will facilitate wider use of EPIC in cognitive science research, as well as easing the process of teaching EPIC to a new generation of modelers in the classroom and in workshops.
 
@@ -28,7 +28,7 @@ Interest in using computational cognitive architectures to understand human cogn
 
 ![Figure 1. Overview of EPICapp Modeling.\label{fig:epicappoverview}](files/epicapp_overview.png)
 
[…]
