### 2022-08-24 #4533 @travisseymour
Prashant,
After reviewing the PDF you created, I realized that there were a couple of
typo fixes and OS version clarifications needed. I made the required
changes and pushed the updates to github.
Please generate a new pdf.
Thanks
Travis

On Wed, Aug 24, 2022 at 8:57 AM The Open Journals editorial robot <
***@***.***> wrote:

> 👉📄 Download article proof
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.04533/joss.04533/10.21105.joss.04533.pdf>
> 📄 View article proof on GitHub
> <https://github.com/openjournals/joss-papers/blob/joss.04533/joss.04533/10.21105.joss.04533.pdf>
> 📄 👈
>
