### 2022-08-25 #4533 @prashjha
Hi @travisseymour, thanks for your response. The draft is well written. Here are a few suggestions on the draft (see if they make sense):

- Remove the extra `Figure 1` from the Figure 1 caption. Same with Figure 2 caption. Also, you may highlight the change in Figure 2 (i.e., python) compared to Figure 1 in the caption of Figure 2.
- I will suggest referring to Figure 1 in line 47: ``of the following components (see Figure 1):``
- Line 67: ``application:``?
- Line 69: ``... with which the author showed ...``?

Thanks for archiving the library. Let me know when you have made the changes (or prefer not to make changes), and I will hand the paper to EiC.
