### 2022-08-25 fce5c5d
corrected spelling of a word and punctuation of another.
