### 2022-08-26 81579aa
user can now specify built-in text editor or point to one on disk.
