+++ src/main/python/mainwindow.py
@@ -21,6 +21,7 @@ from PySide2 import QtWidgets
 import fitness
 from dialogs.aboutwindow import AboutWin
 from dialogs.fontsizewindow import FontSizeDialog
+from dialogs.texteditchoicewindow import TextEditChoiceWin
 from uifiles.mainui import Ui_MainWindow
 from dialogs.sndtextsettingswindow import SoundTextSettingsWin
 from statswindow import StatsWin
@@ -114,7 +115,7 @@ class StateChangeWatcher(QObject):
     """
 
     def __init__(
-        self, parent=None, minimize_func: Callable = None, restore_func: Callable = None
+            self, parent=None, minimize_func: Callable = None, restore_func: Callable = None
     ):
         super(StateChangeWatcher, self).__init__(parent)
         self.minimize_func = minimize_func
@@ -224,6 +225,7 @@ class MainWin(QMainWindow):
         self.epiclib_settings_dialog = None
         self.sound_text_settings_dialog = None
         self.font_size_dialog = None
+        self.text_editor_dialog = None
 
         # setup state change watcher so all min or max together
         # note: this has to go before self.setup_view()
@@ -298,9 +300,9 @@ class MainWin(QMainWindow):
         try:
             self.manage_z_order = False
             for win in chain(
-                [self, self.trace_win, self.stats_win],
-                self.visual_views.values(),
-                self.auditory_views.values(),
+                    [self, self.trace_win, self.stats_win],
+                    self.visual_views.values(),
+                    self.auditory_views.values(),
             ):
                 if win.isVisible():
                     #     or win.windowState() == Qt.WindowMinimized:
@@ -411,6 +413,11 @@ class MainWin(QMainWindow):
         self.ui.actionDark_Mode_Toggle.triggered.connect(self.toggle_darkmode)
         self.ui.actionDark_Mode_Toggle.setChecked(config.app_cfg.dark_mode)
         self.ui.actionDelete_Datafile.triggered.connect(self.delete_datafile)
+
+        self.ui.actionText_Editor.triggered.connect(self.choose_text_editor)
+        te_path = Path(config.app_cfg.text_editor)
+        self.ui.actionText_Editor.setText(f"Text Editor: {te_path.name if te_path.is_file() else 'BUILT-IN'}")
+
     def window_focus_changed(self, win: QMainWindow):
         if not win or not self.manage_z_order:
             return
@@ -734,12 +741,12 @@ class MainWin(QMainWindow):
         )
 
         for win in (
-            self.visual_physical_view,
-            self.visual_sensory_view,
-            self.visual_perceptual_view,
-            self.auditory_physical_view,
-            self.auditory_sensory_view,
-            self.auditory_perceptual_view,
+                self.visual_physical_view,
+                self.visual_sensory_view,
+                self.visual_perceptual_view,
+                self.auditory_physical_view,
+                self.auditory_sensory_view,
+                self.auditory_perceptual_view,
         ):
             win.set_changed()
 
@@ -795,13 +802,13 @@ class MainWin(QMainWindow):
             if self.simulation.tempmod_path and self.simulation.tempmod_path.is_file():
                 self.simulation.tempmod_path.unlink()
             if (
-                self.simulation.tempenc_v_path
-                and self.simulation.tempenc_v_path.is_file()
+                    self.simulation.tempenc_v_path
+                    and self.simulation.tempenc_v_path.is_file()
             ):
                 self.simulation.tempenc_v_path.unlink()
             if (
-                self.simulation.tempenc_a_path
-                and self.simulation.tempenc_a_path.is_file()
+                    self.simulation.tempenc_a_path
[…]
