### 2022-08-26 a16e62c
Right-click on normal output correctly uses specified external editor.
