+++ article/paper.md
@@ -46,7 +46,7 @@ Many studies have been published with EPIC models created using the existing EPI
 
 # Statement of Need
 
-There are several issues that motivate a new approach to EPIC modeling. One is that it is currently not easy to publicly obtain a working copy of the EPICapp software. What is available is the C++ source code for EPICLib and EPICapp for MacOS. This code is available on a Github website [@EPICappGithub], but it contains no information about how to set up or use the required development environment. Furthermore, there is no information on how to program devices, or encoders for use with EPICapp. Without instructions, use of this code would likely be restricted to experienced C++ programmers capable of reading through the source code and extracting the necessary information.
+There are several issues that motivate a new approach to EPIC modeling. One is that it is currently not easy to publicly obtain a working copy of the EPICapp software. What is available is the C++ source code for EPICLib and EPICapp for MacOS. This code is available on a GitHub website [@EPICappGithub], but it contains no information about how to set up or use the required development environment. Furthermore, there is no information on how to program devices, or encoders for use with EPICapp. Without instructions, use of this code would likely be restricted to experienced C++ programmers capable of reading through the source code and extracting the necessary information.
 
 Although there have been a few conference-based workshops given by David Kieras and colleagues [e.g., @Tutorial2004], these have been limited to writing and editing production rules for existing task devices to avoid the need for C++ expertise. Although there are 2 publicly available EPIC models^[https://github.com/dekieras/EPIC_HZ_model and https://github.com/dekieras/EPIC_BMS_model], they are both offered as raw code only. To compile these devices for use in EPICapp, the entire EPICLib/EPICapp development environment would have to be set up. Thus, there is currently no option for non-C++ programmers.
 
@@ -72,7 +72,7 @@ EPICpy includes a large list of improvements and enhancements not available in E
 
 # Limitations
 
-Getting an existing C++ library to work seamlessly with Python code is no easy feat. Although this coordination was partially aided by the cppyy library for the automatic creation of c++/python bindings [@cppyy2016], the back and forth translation between C++ and Python has a significant penalty in run-time. In addition, having the entire user interface, device, and all encoders in Python also contributes to a significant runtime penalty compared to EPICapp. For example, running the provided demo Choice-Task "Hard" condition for 100 trials in EPICpy takes 1.36 seconds to complete compared to 0.81 seconds in EPICapp, making the C++ version 40% faster than the Python version^[These times were achieved by disabling all graphical view output, disabling all text output, and setting all model step delays to 0.0 during each run. Context: Pop!_OS Linux v21.10, AMD Ryzen 7 3700X with 8(16) cores @ 3.6GHz, 64GB RAM]. When continuous graphic and text output are enabled, the speed advantage of C++ is even more dramatic. However, it is our hope that this trades off with the decrease in _development_ time typically associated with moving from C++ to Python, especially for inexperienced developers.
[…]
