### 2022-08-26 #4533 @arfon
@travisseymour – I've made a couple of very small edit here: https://github.com/travisseymour/EPICpy/pull/6

Could you merge that PR please?
