### 2022-08-26 #4533 @travisseymour
How exciting, thanks for all the work done by the reviewers and the
editorial team!

Am I correct that I no longer need the branch used for the review?

On Fri, Aug 26, 2022 at 11:35 AM Arfon Smith ***@***.***>
wrote:

> Closed #4533 <https://github.com/openjournals/joss-reviews/issues/4533>
> as completed.
>
