### 2022-08-27 0818dad
added some citation info now that article is published.
