+++ CITATION.cff
@@ -1,0 +1,11 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as shown below."
+authors:
+- family-names: "Seymour"
+  given-names: "Travis"
+  orcid: "https://orcid.org/0000-0003-4719-9152"
+title: "EPICpy: A Python Interface for EPIC Computational Cognitive Modeling"
+version: 2022.3.5
+doi: 10.5281/zenodo.7020282
+date-released: 2022-10-26
+url: "https://github.com/travisseymour/EPICpy"
