+++ src/main/python/mainwindow.py
@@ -115,7 +115,7 @@ class StateChangeWatcher(QObject):
     """
 
     def __init__(
-            self, parent=None, minimize_func: Callable = None, restore_func: Callable = None
+        self, parent=None, minimize_func: Callable = None, restore_func: Callable = None
     ):
         super(StateChangeWatcher, self).__init__(parent)
         self.minimize_func = minimize_func
@@ -300,9 +300,9 @@ class MainWin(QMainWindow):
         try:
             self.manage_z_order = False
             for win in chain(
-                    [self, self.trace_win, self.stats_win],
-                    self.visual_views.values(),
-                    self.auditory_views.values(),
+                [self, self.trace_win, self.stats_win],
+                self.visual_views.values(),
+                self.auditory_views.values(),
             ):
                 if win.isVisible():
                     #     or win.windowState() == Qt.WindowMinimized:
@@ -416,7 +416,9 @@ class MainWin(QMainWindow):
 
         self.ui.actionText_Editor.triggered.connect(self.choose_text_editor)
         te_path = Path(config.app_cfg.text_editor)
-        self.ui.actionText_Editor.setText(f"Text Editor: {te_path.name if te_path.is_file() else 'BUILT-IN'}")
+        self.ui.actionText_Editor.setText(
+            f"Text Editor: {te_path.name if te_path.is_file() else 'BUILT-IN'}"
+        )
 
     def window_focus_changed(self, win: QMainWindow):
         if not win or not self.manage_z_order:
@@ -741,12 +743,12 @@ class MainWin(QMainWindow):
         )
 
         for win in (
-                self.visual_physical_view,
-                self.visual_sensory_view,
-                self.visual_perceptual_view,
-                self.auditory_physical_view,
-                self.auditory_sensory_view,
-                self.auditory_perceptual_view,
+            self.visual_physical_view,
+            self.visual_sensory_view,
+            self.visual_perceptual_view,
+            self.auditory_physical_view,
+            self.auditory_sensory_view,
+            self.auditory_perceptual_view,
         ):
             win.set_changed()
 
@@ -802,13 +804,13 @@ class MainWin(QMainWindow):
             if self.simulation.tempmod_path and self.simulation.tempmod_path.is_file():
                 self.simulation.tempmod_path.unlink()
             if (
-                    self.simulation.tempenc_v_path
-                    and self.simulation.tempenc_v_path.is_file()
+                self.simulation.tempenc_v_path
+                and self.simulation.tempenc_v_path.is_file()
             ):
                 self.simulation.tempenc_v_path.unlink()
             if (
-                    self.simulation.tempenc_a_path
-                    and self.simulation.tempenc_a_path.is_file()
+                self.simulation.tempenc_a_path
+                and self.simulation.tempenc_a_path.is_file()
             ):
                 self.simulation.tempenc_a_path.unlink()
 
@@ -838,7 +840,7 @@ class MainWin(QMainWindow):
             self.manage_z_order = True
 
     def set_view_background_image(
-            self, view_type: str, img_filename: str, scaled: bool = True
+        self, view_type: str, img_filename: str, scaled: bool = True
     ):
         if not config.device_cfg.allow_device_images:
             return
@@ -878,12 +880,12 @@ class MainWin(QMainWindow):
         self.trace_win.ui.plainTextEditOutput.cache_text = not enable
 
     def clear_ui(
-            self,
-            visual_views: bool = True,
-            auditory_views: bool = True,
-            normal_output: bool = True,
-            trace_output: bool = True,
-            stats_output: bool = True,
+        self,
+        visual_views: bool = True,
+        auditory_views: bool = True,
+        normal_output: bool = True,
+        trace_output: bool = True,
+        stats_output: bool = True,
     ):
         if visual_views:
[…]
