### 2022-08-28 a72b12d
fixed typo in EVERY instance of short license statement.
