+++ src/main/python/mainwindow.py
@@ -9,7 +9,7 @@ EPICpy is distributed in the hope that it will be useful, but WITHOUT ANY WARRAN
 without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 See the GNU General Public License for more details.
 
-You should have received a copy of the GNU General Public License along with Foobar.
+You should have received a copy of the GNU General Public License along with EPICpy.
 If not, see <https://www.gnu.org/licenses/>.
 """
 import os
