+++ src/main/python/uifiles/aboutui.py
@@ -105,7 +105,7 @@ class Ui_aboutDialog(object):
                 "<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-family:'JetBrains Mono','monospace'; font-size:15.8pt; color:#000000;\">of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.</span></p>\n"
                 '<p style="-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-'
                 "indent:0; text-indent:0px; font-family:'JetBrains Mono','monospace'; font-size:15.8pt; color:#000000;\"><br /></p>\n"
-                "<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-family:'JetBrains Mono','monospace'; font-size:15.8pt; color:#000000;\">You should have received a copy of the GNU General Public License along with Foobar.</span></p>\n"
+                "<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-family:'JetBrains Mono','monospace'; font-size:15.8pt; color:#000000;\">You should have received a copy of the GNU General Public License along with EPICpy.</span></p>\n"
                 "<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-family:'JetBrains Mono','monospace'; font-size:15.8pt; color:#000000;\">If not, see &lt;https://www.gnu.org/licenses/&gt;.</span></p>\n"
                 "<p style=\"-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px; font-family:'JetBrains Mono','monospace'; font-size:13.5pt; color:#808080;\"><br /></p>\n"
                 '<p style=" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0'
