+++ src/main/python/version.py
@@ -14,7 +14,7 @@ If not, see <https://www.gnu.org/licenses/>.
 """
 
 # year.month.version
-__version__ = "2022.4.2"
+__version__ = "2022.10.1"
 
 __loglevel__ = "INFO"
 
