+++ src/main/python/epicsimulation.py
@@ -20,8 +20,8 @@ import sys
 import weakref
 import timeit
 
-from PySide2.QtCore import QTimer
-from PySide2.QtWidgets import QFileDialog
+from PyQt5.QtCore import QTimer
+from PyQt5.QtWidgets import QFileDialog
 
 from apputils import LIBNAME, unpack_param_string
 from cppinclude import epiclib_include
