### 2022-08-28 a1b554e
changed: EPICpy license changed to GPLv3 due to use of PyQT5.
