+++ src/main/python/dialogs/tracesettingswindow.py
@@ -1,16 +1,21 @@
 """
-This file is part of EPICpy, Created by Travis L. Seymour, PhD.
+This file is part of the EPICpy source code. EPICpy is a tool for simulating 
+human performance tasks using the EPIC computational cognitive architecture 
+(David Kieras and David Meyer 1997a) using the Python programming language.
+Copyright (C) 2022 Travis L. Seymour, PhD
 
-EPICpy is free software: you can redistribute it and/or modify it under the terms of
-the GNU Lesser General Public License as published by the Free Software Foundation,
-either version 3 of the License, or (at your option) any later version.
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
 
-EPICpy is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
-without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-See the GNU General Public License for more details.
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
 
-You should have received a copy of the GNU General Public License along with EPICpy.
-If not, see <https://www.gnu.org/licenses/>.
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
 from uifiles.tracesettingsui import Ui_DialogTraceSettings
