+++ src/main/python/epiccoder/mainwindow.py
@@ -1,0 +1,655 @@
+"""
+EPICcoder is a minimal programmer's text editor created for use with the
+EPIC, EPICpy and pyEPIC simulation environments.
+Copyright (C) 2022 Travis L. Seymour, PhD
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""
+
+import os
+from typing import Optional
+
+# from PyQt5.QtWidgets import *
+# from PyQt5.QtCore import *
+# from PyQt5.QtGui import *
+
+import sys
+from pathlib import Path
+
+from PyQt5.QtCore import QSettings, Qt, QSize, QDir, QModelIndex
+from PyQt5.QtGui import QIcon, QFont, QPixmap, QResizeEvent, QCloseEvent
+from PyQt5.QtWidgets import (
+    QMainWindow,
+    QLabel,
+    QFrame,
+    QSizePolicy,
+    QHBoxLayout,
+    QVBoxLayout,
+    QSplitter,
+    QFileSystemModel,
+    QTreeView,
+    QLineEdit,
+    QCheckBox,
+    QListWidget,
+    QSpacerItem,
+    QTabWidget,
+    QMessageBox,
+    QFileDialog,
+    QApplication,
+)
+
+from epiccoder.customeditor import CustomEditor
+from epiccoder.fuzzy_searcher import SearchItem, SearchWorker
+from epiccoder.aboutwindow import AboutWin
+from epiccoder.version import __version__
+
+from fbs_runtime.application_context.PyQt5 import ApplicationContext
+
+int_gen = (
+    i for i in range(sys.maxsize)
+)  # generator for successive integers from 1 to sys.maxsize
+
+
+class MainEditorWindow(QMainWindow):
+    def __init__(self, context: ApplicationContext):
+        super(QMainWindow, self).__init__()
+
+        self.context = context
+
+        # add before init
+        self.side_bar_clr = "#282c34"
+
+        # init settings
+        self.settings = QSettings()
+        self.main_window_width = self.settings.value(
+            "main_window_width", type=int, defaultValue=1300
+        )
+        self.main_window_height = self.settings.value(
+            "main_window_height", type=int, defaultValue=900
+        )
+
+        self.init_ui()
+
+        self.current_file = None
+        self.current_side_bar = None
+
+        self.setWindowIcon(QIcon(self.context.get_resource("uiicons", "app-icon.svg")))
+
+    def init_ui(self):
+        self.setWindowTitle(f"EPIC Coder v{__version__}")
+        self.resize(self.main_window_width, self.main_window_height)
+
+        style_sheet = Path(self.context.get_resource("css/style.qss")).read_text()
+        style_sheet = style_sheet.replace(
+            ":/icons/close-icon.svg",
+            self.context.get_resource("uiicons", "close-icon.svg"),
+        )
+        self.setStyleSheet(style_sheet)
+
+        # setup font
+        self.window_font = QFont("Fira Code")
+        self.window_font.setPointSize(12)
+        self.setFont(self.window_font)
+
+        self.set_up_menu()
+        self.set_up_body()
+        self.statusBar().showMessage("hello")
+
+        self.show()
+
+    def set_up_menu(self):
+        menu_bar = self.menuBar()
+
+        # File Menu
+        file_menu = menu_bar.addMenu("File")
+
[…]
