+++ src/main/resources/base/uiicons/NOTE.txt
@@ -1,0 +1,1 @@
+these icons are for use with the built-in editor (based on EPICcoder)
