+++ requirements.txt
@@ -4,6 +4,7 @@
 loguru
 fbs  # ok to run/develop EPICpy, but to use any fbs ____ command on Py3.7+, you'll need fbs pro
 PyQt5
+QScintilla
 python-dateutil
 ulid
 pingouin

[…]
