### 2022-08-28 a9aa31f
fixed: editor won't close automatically if any open file has changed.
