+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 
 # EPICpy Changelog
-#### (lasted updated Sun Aug 28 13:54:01 2022)
+#### (lasted updated Sun Aug 28 15:30:36 2022)
 
 ---
 
-#### 1. CHANGE: EPICpy license changed to GPLv3 due to use of PyQT5.
+#### 1. NEW FEATURE: Built-in editor is available (based on the EPICcoder project)
 
-#### 2. CHANGE: EPICpy GUI qt-bindings switched from Pyside2 to PyQt5
+#### 2. CHANGE: EPICpy license changed to GPLv3 due to use of PyQT5.
+
+#### 3. CHANGE: EPICpy GUI qt-bindings switched from Pyside2 to PyQt5
 
 
 ## VERSION: 2022.10.1
