+++ src/main/python/version.py
@@ -19,7 +19,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
 # year.month.version
-__version__ = "2022.10.1"
+__version__ = "2022.8.2"
 
 __loglevel__ = "INFO"
 
