### 2022-08-29 3cb20d9
exact_requirements.txt was getting out of date (e.g., we're now using PyQt5)
