+++ exact_requirements.txt
@@ -1,18 +1,32 @@
 altgraph==0.17.2
+ascii-frame==0.0.1
 asttokens==2.0.5
+attrs==21.4.0
+bcrypt==3.2.0
+black==22.1.0
+boto3==1.21.21
+botocore==1.24.21
+bullet==2.2.0
 certifi==2021.10.8
+cffi==1.15.0
+chardet==4.0.0
 charset-normalizer==2.0.12
 click==8.0.4
 colorama==0.4.4
+cryptography==36.0.2
 cycler==0.11.0
+distro==1.7.0
+evdev==1.5.0
 executing==0.8.3
-fbs==1.1.0
 fonttools==4.29.1
+future==0.18.2
 ghp-import==2.0.2
 icecream==2.1.2
 idna==3.3
 importlib-metadata==4.11.2
+iniconfig==1.1.1
 Jinja2==3.0.3
+jmespath==1.0.0
 joblib==1.1.0
 kiwisolver==1.3.2
 littleutils==0.2.2
@@ -22,24 +36,44 @@ MarkupSafe==2.1.0
 matplotlib==3.5.1
 mergedeep==1.3.4
 mkdocs==1.2.3
+mypy-extensions==0.4.3
 numpy==1.22.3
 outdated==0.2.1
 packaging==21.3
 pandas==1.4.1
 pandas-flavor==0.2.0
+paramiko==2.10.2
+pathspec==0.9.0
 patsy==0.5.2
 Pillow==9.0.1
 pingouin==0.5.1
+platformdirs==2.5.1
+pluggy==1.0.0
+plum-dispatch==1.5.8
+plumbum==1.7.2
+py==1.11.0
+pycparser==2.21
 Pygments==2.11.2
 pyinstaller==4.10
 pyinstaller-hooks-contrib==2022.2
+PyNaCl==1.5.0
+pynput==1.7.6
 pyparsing==3.0.7
-PySide2==5.15.2.1
+PyQt5==5.15.7
+PyQt5-Qt5==5.15.2
+PyQt5-sip==12.11.0
+PyQt5-stubs==5.15.6.0
+pysftp==0.2.9
+pytest==7.1.1
+pytest-qt==4.0.2
 python-dateutil==2.8.2
+python-xlib==0.31
 pytz==2021.3
 PyYAML==6.0
 pyyaml_env_tag==0.1
+QScintilla==2.13.3
 requests==2.27.1
+s3transfer==0.5.2
 scipy==1.8.0
 seaborn==0.11.2
 shiboken2==5.15.2.1
@@ -47,6 +81,8 @@ six==1.16.0
 statsmodels==0.13.2
 tabulate==0.8.9
 threadpoolctl==3.1.0
+tomli==2.0.1
+typing_extensions==4.1.1
 ulid==1.1
 urllib3==1.26.8
 watchdog==2.1.6
