### 2022-08-29 029e91e
removed CITATION.cff in favor of explicit citation note in README.md
